+++ server/index.js
@@ -1,13 +1,11 @@
 import express from 'express';
-import mongoose from 'mongoose';
+import './db/db.js'; 
 import cors from 'cors';
 import dotenv from 'dotenv';
 
 import userRoutes from './routes/userRoutes.js';
 import promptRoutes from './routes/promptRoutes.js';
 import gameRoutes from './routes/gameRoutes.js';
-import statsRoutes from './routes/statRoutes.js';
-// import projectRoutes from './routes/projectRoutes.js';
 
 dotenv.config();
 
@@ -20,7 +18,5 @@ app.use(express.json());
 app.use('/users', userRoutes);
 app.use('/prompts', promptRoutes);
 app.use('/games', gameRoutes);
-app.use('/stats', statsRoutes);
-// app.use('/projects', projectRoutes);
 
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
