+++ server/models/gameSchema.js
@@ -1,12 +1,35 @@
 import mongoose from 'mongoose';
 
 const gameSchema = new mongoose.Schema({
+    user_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true }, 
+//   prompt_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Prompt', required: true }, //make array
+  prompts: [
+  {
+    prompt_id: {
+      type: mongoose.Schema.Types.ObjectId,
+      required: true,
+      ref: 'Prompt',
+    },
+  },
+],
   title: { type: String, required: true },
   description: { type: String, required: true },
-  user_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true },
-  week: { type: Number, default: 0 }, //this should take from the Stats week field, or maybe we could just write to both at the same time? idk
+  week: { type: Number, default: 0, required: true },
+  discovery: { type: String },
+  discussion: { type: String },
+  abundance: { type: String },
+  scarcity: { type: String },
+  contempt: { type: Number, default: 0 },
+  project_title: { type: String },
+  project_desc: { type: String },
+  project_weeks: { type: Number, min: 0, max: 6 },
+  pp_title: { type: String },
+  pp_desc: { type: String },
+  pp_weeks: { type: Number, min: 0, max: 6 },
+  p_discussion: { type: String },
+  p_discovery: { type: String },
   end: { type: String }
 });
 
-const Game = mongoose.model('Game', gameSchema);
+const Game = mongoose.model('Stats', gameSchema);
 export default Game;
