--- server/routes/projectRoutes.js
@@ -1,18 +1,0 @@
-// import express from 'express';
-// import {
-//   getAllProjects,
-//   createProject,
-//   getProjectById,
-//   updateProject,
-//   deleteProject,
-// } from '../controllers/projectController.js';
-
-// const router = express.Router();
-
-// router.get('/', getAllProjects);
-// router.post('/', createProject);
-// router.get('/:id', getProjectById);
-// router.put('/:id', updateProject);
-// router.delete('/:id', deleteProject);
-
-// export default router;
